Add MaterializedResult projetion methods
Move them from `QueryAssert` to `MaterializedResult`.
