+++ core/trino-main/src/test/java/io/trino/sql/query/QueryAssertions.java
@@ -15,7 +15,6 @@ package io.trino.sql.query;
 
 import com.google.common.base.Joiner;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Lists;
 import io.trino.Session;
 import io.trino.execution.warnings.WarningCollector;
@@ -52,15 +51,11 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
-import java.util.Set;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
-import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.collect.ImmutableList.toImmutableList;
 import static io.airlift.testing.Assertions.assertEqualsIgnoreOrder;
 import static io.trino.cost.StatsCalculator.noopStatsCalculator;
 import static io.trino.execution.querystats.PlanOptimizersStatsCollector.createPlanOptimizersStatsCollector;
@@ -329,55 +324,26 @@ public class QueryAssertions
 
         public QueryAssert exceptColumns(String... columnNamesToExclude)
         {
-            validateIfColumnsPresent(columnNamesToExclude);
-            checkArgument(columnNamesToExclude.length > 0, "At least one column must be excluded");
-            checkArgument(columnNamesToExclude.length < actual.getColumnNames().size(), "All columns cannot be excluded");
-            return projected(((Predicate<String>) Set.of(columnNamesToExclude)::contains).negate());
-        }
-
-        public QueryAssert projected(String... columnNamesToInclude)
-        {
-            validateIfColumnsPresent(columnNamesToInclude);
-            checkArgument(columnNamesToInclude.length > 0, "At least one column must be projected");
-            return projected(Set.of(columnNamesToInclude)::contains);
-        }
-
-        private QueryAssert projected(Predicate<String> columnFilter)
-        {
-            List<String> columnNames = actual.getColumnNames();
-            Map<Integer, String> columnsIndexToNameMap = new HashMap<>();
-            for (int i = 0; i < columnNames.size(); i++) {
-                String columnName = columnNames.get(i);
-                if (columnFilter.test(columnName)) {
-                    columnsIndexToNameMap.put(i, columnName);
-                }
-            }
-
             return new QueryAssert(
                     runner,
                     session,
-                    format("%s projected with %s", query, columnsIndexToNameMap.values()),
-                    new MaterializedResult(
-                            actual.getMaterializedRows().stream()
-                                    .map(row -> new MaterializedRow(
-                                            row.getPrecision(),
-                                            columnsIndexToNameMap.keySet().stream()
-                                                    .map(row::getField)
-                                                    .collect(toList()))) // values are nullable
-                                    .collect(toImmutableList()),
-                            columnsIndexToNameMap.keySet().stream()
-                                    .map(actual.getTypes()::get)
-                                    .collect(toImmutableList())),
+                    format("%s except columns %s", query, Arrays.toString(columnNamesToExclude)),
+                    actual.exceptColumns(columnNamesToExclude),
                     ordered,
                     skipTypesCheck,
                     skipResultsCorrectnessCheckForPushdown);
         }
 
-        private void validateIfColumnsPresent(String... columns)
+        public QueryAssert projected(String... columnNamesToInclude)
         {
-            Set<String> columnNames = ImmutableSet.copyOf(actual.getColumnNames());
-            Arrays.stream(columns)
-                    .forEach(column -> checkArgument(columnNames.contains(column), "[%s] column is not present in %s".formatted(column, columnNames)));
+            return new QueryAssert(
+                    runner,
+                    session,
+                    format("%s projected with %s", query, Arrays.toString(columnNamesToInclude)),
+                    actual.project(columnNamesToInclude),
+                    ordered,
+                    skipTypesCheck,
+                    skipResultsCorrectnessCheckForPushdown);
         }
 
         public QueryAssert matches(BiFunction<Session, QueryRunner, MaterializedResult> evaluator)
